Refactoring: Isolated worker types in their own module and use lenses for accessing their values
